split stuff up
