+++ symbolic-il2cpp/src/binary/mod.rs
@@ -1,0 +1,5 @@
+mod dwarf;
+mod executable;
+
+pub use dwarf::DwarfData;
+pub use executable::Il2CppCodeGenModule;
