+++ symbolic-il2cpp/src/utils.rs
@@ -1,0 +1,1 @@
+pub const CStrCtx: scroll::ctx::StrCtx = scroll::ctx::StrCtx::Delimiter(scroll::ctx::NULL);
